Rename files extension to .mts
