+++ packages/core/src/lib/prefix.mts
@@ -1,6 +1,6 @@
 import chalk from 'chalk';
 import spinners from 'cli-spinners';
-import { useState, useEffect } from '../index.js';
+import { useState, useEffect } from '../index.mjs';
 
 const spinner = spinners.dots;
